+++ interface.go
@@ -1,5 +1,5 @@
 package nnfcgolang
 
 type nn interface {
-	Predict(data [][]float32) ([][]float32, error)
+	Predict(data [][]float64) ([][]float64, error)
 }
